1.1.0+alpha.1 (#50)

+++ core/src/main/java/com/zigythebird/playeranimcore/animation/layered/modifier/AbstractFadeModifier.java
@@ -41,7 +41,7 @@ import org.jetbrains.annotations.Nullable;
  */
 public abstract class AbstractFadeModifier extends AbstractModifier {
     protected int length;
-    protected float time;
+    protected int time;
     protected float tickDelta;
 
     public void setTransitionAnimation(@Nullable IAnimation transitionAnimation) {
